+++ CHANGELOG.md
@@ -10,4 +10,4 @@
 
 ###### Nov 02, 2021
 - Update .env Backend endpoint
-- Testing123
+- Change logout method